Split Firestore into Expense and Payment services
