+++ lib/ui/widgets/dialogs/assignee_picker_dialog.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:statera/data/models/expense.dart';
 import 'package:statera/data/models/group.dart';
+import 'package:statera/data/services/expense_service.dart';
 import 'package:statera/data/services/firestore.dart';
 import 'package:statera/ui/widgets/author_avatar.dart';
 import 'package:statera/ui/widgets/custom_stream_builder.dart';
@@ -70,7 +71,7 @@ class _AssigneePickerDialogState extends State<AssigneePickerDialog> {
         ElevatedButton(
           onPressed: () async {
             widget.expense.updateAssignees(this._selectedUids);
-            await Firestore.instance.updateExpense(widget.expense);
+            await ExpenseService.updateExpense(widget.expense);
             Navigator.pop(context);
           },
           child: Text('Save'),
